signup flow integrated with dashboard

+++ app/src/pages/api/login/index.ts
@@ -5,7 +5,16 @@ import { prisma } from "@/lib/db"
 
 
 async function GET(req:NextApiRequest,res:NextApiResponse){
-    const users = await prisma.user.deleteMany();
+    
+    const {action} = req.query;
+    let users;
+
+    if (action==="delete"){
+      users = await prisma.user.deleteMany();
+    }
+    else{
+      users = await prisma.user.findMany();
+    }
     res.status(200).json(users);
 }
